requirements.txt
Flask
python-pptx
Pillow
openpyxl
